feat(lab1): Изменить под новый вариант

+++ src/main/java/com/enifl33fi/lab1/api/model/product/Coordinates.java
@@ -12,8 +12,8 @@ import lombok.*;
 @EqualsAndHashCode(callSuper = true)
 public class Coordinates extends OwnedEntity {
   @Column(nullable = false)
-  private int x;
+  private Float x;
 
   @Column(nullable = false)
-  private Float y;
+  private Double y;
 }
